:technologist: Convert clean script to a class
